feature: add 32 byte support for rangedb

+++ furano-common/src/main/java/com/cryptoeconomicslab/furano_common/db/range/RangeStore.kt
@@ -1,10 +1,11 @@
 package com.cryptoeconomicslab.furano_common.db.range
 
 import com.cryptoeconomicslab.furano_common.types.Bytes
+import java.math.BigInteger
 
 interface RangeStore {
-    fun get(start: Long, end: Long): Array<RangeRecord>
-    fun put(start: Long, end: Long, value: Bytes)
-    fun del(start: Long, end: Long)
+    fun get(start: BigInteger, end: BigInteger): Array<RangeRecord>
+    fun put(start: BigInteger, end: BigInteger, value: Bytes)
+    fun del(start: BigInteger, end: BigInteger)
     fun bucket(key: Bytes): RangeStore
 }
